+++ crates/feos-core/src/equation_of_state/residual.rs
@@ -3,9 +3,10 @@ use nalgebra::{DVector, DefaultAllocator, Dim, Dyn, OMatrix, OVector, U1, alloca
 use num_dual::{DualNum, Gradients, partial, partial2, second_derivative, third_derivative};
 use quantity::ad::first_derivative;
 use quantity::*;
-use std::ops::Deref;
+use std::ops::{Deref, Div};
 use std::sync::Arc;
-use typenum::Quot;
+
+type Quot<T1, T2> = <T1 as Div<T2>>::Output;
 
 /// Molar weight of all components.
 ///