+++ crates/feos/tests/gc_pcsaft/dft.rs
@@ -10,7 +10,6 @@ use feos_dft::{DFTSolver, Geometry};
 use nalgebra::dvector;
 use quantity::*;
 use std::error::Error;
-use typenum::P3;
 
 #[test]
 #[allow(non_snake_case)]
@@ -40,7 +39,7 @@ fn test_bulk_implementation() -> Result<(), Box<dyn Error>> {
     let eos = GcPcSaft::new(parameters);
     let func = GcPcSaftFunctional::new(parameters_func);
     let t = 200.0 * KELVIN;
-    let v = 0.002 * METER.powi::<P3>() * NAV / NAV_old;
+    let v = 0.002 * METER.powi::<3>() * NAV / NAV_old;
     let n = dvector![1.5] * MOL;
     let state_eos = State::new_nvt(&&eos, t, v, &n)?;
     let state_func = State::new_nvt(&&func, t, v, &n)?;
@@ -123,7 +122,7 @@ fn test_bulk_association() -> Result<(), Box<dyn Error>> {
     let func = GcPcSaftFunctional::new(func_parameters);
 
     let t = 200.0 * KELVIN;
-    let v = 0.002 * METER.powi::<P3>();
+    let v = 0.002 * METER.powi::<3>();
     let n = dvector![1.5] * MOL;
     let state_eos = State::new_nvt(&&eos, t, v, &n)?;
     let state_func = State::new_nvt(&&func, t, v, &n)?;
@@ -171,13 +170,13 @@ fn test_dft() -> Result<(), Box<dyn Error>> {
 
     assert_relative_eq!(
         vle.vapor().density,
-        12.8820179191167643 * MOL / METER.powi::<P3>() * NAV_old / NAV,
+        12.8820179191167643 * MOL / METER.powi::<3>() * NAV_old / NAV,
         max_relative = 1e-13,
     );
 
     assert_relative_eq!(
         vle.liquid().density,
-        13.2705903446123212 * KILO * MOL / METER.powi::<P3>() * NAV_old / NAV,
+        13.2705903446123212 * KILO * MOL / METER.powi::<3>() * NAV_old / NAV,
         max_relative = 1e-13,
     );